feat(pkg): add non-privileged audit command to look for unexpected manually installed packages

+++ bin/add-deb-packages.sh
@@ -2,7 +2,7 @@
 set -e
 LEVEL="${1:-base}"
 
-if [ "$(id -u)" -ne '0' ]; then
+if [ "$LEVEL" != "audit" ] && [ "$(id -u)" -ne '0' ]; then
     if [ -z "$NO_RECURSE" ]; then
         exec sudo NO_RECURSE=1 /bin/bash -e "$0" "$LEVEL"
     else
@@ -16,7 +16,9 @@ apt-config dump | grep -E 'APT::Install' | sed -re 's/1/0/g' > "$TEMPFILE"
 export APT_CONFIG="$TEMPFILE"
 export DEBIAN_FRONTEND=noninteractive
 
-apt update
+if [ "$LEVEL" != "audit" ]; then
+    apt update
+fi
 
 case "$LEVEL" in
     "base")
@@ -48,6 +50,33 @@ case "$LEVEL" in
             fbset ~n^mesa xdg-utils x11-utils x11-xserver-utils git-filter-repo libsecret-tools \
             shellcheck
     ;;
+    "audit")
+        EXPLICIT_LIST=(
+            aptitude aptitude-doc-en bash-completion vim-nox git rsync pinentry-tty gpg-agent
+            bsd-mailx exim4-daemon-light patch zip unzip jq plocate neovim restic curl openssl
+            bsdutils ncal rfkill wpasupplicant w3m parted bc dc kmod btrfs-progs tcpdump wget
+            wodim busybox-static pinentry-fltk pass dbus-user-session ripgrep tree yq
+            apt-file arch-test autoconf automake autotools-dev build-essential debhelper
+            debian-keyring debootstrap devscripts dh-make dkms dosfstools dpkg-dev dput
+            dupload e2fsprogs-l10n eatmydata equivs fakeroot fancontrol gdisk gnupg hdparm
+            htop i2c-tools irqbalance lintian shared-mime-info xauth linux-headers-amd64
+            lm-sensors localepurge manpages-dev mutt net-tools nocache nvme-cli patchutils
+            pbuilder pigz powermgmt-base read-edid screen smartmontools strace
+            thin-provisioning-tools xutils-dev xdg-user-dirs gdb linux-doc info iw bison
+            flex libncurses-dev libelf-dev libssl-dev zstd cpio dwarves xsel upower alsa-utils
+            debconf-utils eject ethtool packagekit cifs-utils vdpau-driver-all va-driver-all
+            exfatprogs exfat-fuse fbset xdg-utils x11-utils x11-xserver-utils git-filter-repo
+            libsecret-tools shellcheck
+        )
+        manual_pkgs="$(aptitude search -F '%p' '?and(?installed, ?not(?automatic), ?not(~slibs), ?not(~v))' | sort -u)"
+        priority_pkgs="$(aptitude search -F '%p' '?and(?installed, ?or(~prequired,~pimportant,~pstandard))' | sort -u)"
+        pattern_pkgs="$(aptitude search -F '%p' '?and(?installed, ?or(~n^plymouth_, ~n^mesa))' | sort -u)"
+        explicit_pkgs="$(printf '%s\n' "${EXPLICIT_LIST[@]}" | sort -u)"
+
+        allowed_pkgs="$(printf '%s\n' "$priority_pkgs" "$pattern_pkgs" "$explicit_pkgs" | sort -u)"
+
+        comm -23 <(echo "$manual_pkgs") <(echo "$allowed_pkgs")
+    ;;
     "cleanup")
         dpkg -l | awk 'c&&!/ii/{print $2}/^\+/{c=1}' | xargs aptitude purge -y
     ;;
@@ -63,6 +92,8 @@ case "$LEVEL" in
     exit 1
 esac
 
-apt autoremove
+if [ "$LEVEL" != "audit" ]; then
+    apt autoremove
+fi
 
 exit 0
